reformat imports, move stream
